split out arena
